+++ src/full/Agda/Main.hs
@@ -87,7 +87,7 @@ runAgda' backends = runTCMPrettyErrors $ do
 
       case mode of
         MainModePrintHelp hp   -> liftIO $ printUsage bs hp
-        MainModePrintVersion   -> liftIO $ printVersion bs
+        MainModePrintVersion o -> liftIO $ printVersion bs o
         MainModePrintAgdaDir   -> liftIO $ printAgdaDir
         MainModeRun interactor -> do
           setTCLens stBackends bs
@@ -97,17 +97,17 @@ runAgda' backends = runTCMPrettyErrors $ do
 data MainMode
   = MainModeRun (Interactor ())
   | MainModePrintHelp Help
-  | MainModePrintVersion
+  | MainModePrintVersion PrintAgdaVersion
   | MainModePrintAgdaDir
 
 -- | Determine the main execution mode to run, based on the configured backends and command line options.
 -- | This is pure.
 getMainMode :: MonadError String m => [Backend] -> Maybe AbsolutePath -> CommandLineOptions -> m MainMode
 getMainMode configuredBackends maybeInputFile opts
-  | Just hp <- optPrintHelp opts = return $ MainModePrintHelp hp
-  | optPrintVersion opts         = return $ MainModePrintVersion
-  | optPrintAgdaDir opts         = return $ MainModePrintAgdaDir
-  | otherwise                    = do
+  | Just hp <- optPrintHelp opts    = return $ MainModePrintHelp hp
+  | Just o  <- optPrintVersion opts = return $ MainModePrintVersion o
+  | optPrintAgdaDir opts            = return $ MainModePrintAgdaDir
+  | otherwise = do
       mi <- getInteractor configuredBackends maybeInputFile opts
       -- If there was no selection whatsoever (e.g. just invoked "agda"), we just show help and exit.
       return $ maybe (MainModePrintHelp GeneralHelp) MainModeRun mi
@@ -257,8 +257,9 @@ backendUsage (Backend b) =
     map void (commandLineFlags b)
 
 -- | Print version information.
-printVersion :: [Backend] -> IO ()
-printVersion backends = do
+printVersion :: [Backend] -> PrintAgdaVersion -> IO ()
+printVersion _ PrintAgdaNumericVersion = putStrLn versionWithCommitInfo
+printVersion backends PrintAgdaVersion = do
   putStrLn $ "Agda version " ++ versionWithCommitInfo
   mapM_ putStrLn
     [ "  - " ++ name ++ " backend version " ++ ver